+++ src/table_description.rs
@@ -162,7 +162,7 @@ impl Drop for TableDescription {
     }
 }
 
-#[cfg(all(test, feature = "bundled-test"))]
+#[cfg(all(test, feature = "_duckdb-testing"))]
 mod tests {
     use super::*;